+++ .github/CHANGELOG.md
@@ -6,9 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 (Dates are in YYYY-MM-DD format. This message is mainly for my own sake.)
 
 ## [Unreleased]
-
 ### Added
 * `<p>restart` command
+* `<p>goroutinestack` goroutine stack dump command
 
 ## [4.8.5] - 2021-07-28
 ### Changed